refactor(mobile): isolate history models from transaction screens
